# architecture/API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: YouTube Niche Discovery Engine API
  description: |
    A comprehensive API for discovering, validating, and analyzing YouTube niches.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Standard users: 100 requests per hour
    - Premium users: 1000 requests per hour
    - Rate limit headers are included in all responses
    
    ## Response Format
    All responses follow a consistent format with `data`, `meta`, and optional `errors` fields.
    
    ## Pagination
    List endpoints support cursor-based pagination for optimal performance.
  version: 1.0.0
  contact:
    name: API Support
    email: support@nichediscovery.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.nichediscovery.com/v1
    description: Production server
  - url: https://staging-api.nichediscovery.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /niches:
    get:
      tags: [Niches]
      summary: List niches with filtering and pagination
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of niches to return
        - name: cursor
          in: query
          schema:
            type: string
          description: Cursor for pagination
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category slug
        - name: min_score
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
          description: Minimum overall score
        - name: max_score
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
          description: Maximum overall score
        - name: status
          in: query
          schema:
            type: string
            enum: [discovered, analyzing, validated, rejected, archived]
          description: Filter by niche status
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [score, discovered_at, name, updated_at]
            default: score
          description: Sort field
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
        - name: search
          in: query
          schema:
            type: string
          description: Full-text search query
        - name: keywords
          in: query
          schema:
            type: string
          description: Comma-separated keywords to filter by
      responses:
        '200':
          description: List of niches
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NicheSummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Niches]
      summary: Manually add a new niche
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NicheCreate'
      responses:
        '201':
          description: Niche created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/NicheDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /niches/{id}:
    get:
      tags: [Niches]
      summary: Get detailed niche information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Niche ID
      responses:
        '200':
          description: Niche details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/NicheDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Niches]
      summary: Update niche information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Niche ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NicheUpdate'
      responses:
        '200':
          description: Niche updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/NicheDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Niches]
      summary: Archive a niche
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Niche ID
      responses:
        '204':
          description: Niche archived successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - code: "INVALID_PARAMETER"
                message: "The 'min_score' parameter must be between 0 and 100"
                field: "min_score"
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - code: "UNAUTHORIZED"
                message: "Invalid or expired authentication token"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - code: "NOT_FOUND"
                message: "Niche with ID '123e4567-e89b-12d3-a456-426614174000' not found"
    Conflict:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    NicheSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "AI-Powered Fitness Apps"
        slug:
          type: string
          example: "ai-powered-fitness-apps"
        overall_score:
          type: number
          minimum: 0
          maximum: 100
          example: 87.5
        status:
          type: string
          enum: [discovered, analyzing, validated, rejected, archived]
          example: "validated"
        category:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
              example: "Health & Fitness"
            slug:
              type: string
              example: "health-fitness"
        discovered_at:
          type: string
          format: date-time
          example: "2026-02-01T10:30:00Z"
        last_analyzed:
          type: string
          format: date-time
          example: "2026-02-01T14:15:00Z"
        keywords:
          type: array
          items:
            type: string
          example: ["AI", "fitness", "apps", "machine learning", "health"]
        tags:
          type: array
          items:
            type: string
          example: ["trending", "high-potential", "tech"]
    NicheDetail:
      allOf:
        - $ref: '#/components/schemas/NicheSummary'
        - type: object
          properties:
            description:
              type: string
              example: "Niche focusing on AI-powered fitness applications that provide personalized workout plans and nutrition advice."
            competition_score:
              type: number
              minimum: 0
              maximum: 100
              example: 65.0
            demand_score:
              type: number
              minimum: 0
              maximum: 100
              example: 92.0
            monetization_score:
              type: number
              minimum: 0
              maximum: 100
              example: 88.0
            trend_score:
              type: number
              minimum: 0
              maximum: 100
              example: 95.0
            discovery_source:
              type: string
              example: "youtube"
            validation_status:
              type: string
              enum: [pending, in_progress, completed, failed]
              example: "completed"
            metadata:
              type: object
              additionalProperties: true
              example:
                primary_platform: "youtube"
                estimated_monthly_searches: 125000
                top_competitors: ["Fitness Blender", "Nike Training Club"]
    NicheCreate:
      type: object
      required: [name, category_id]
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
        category_id:
          type: string
          format: uuid
        keywords:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
    NicheUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
        category_id:
          type: string
          format: uuid
        keywords:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [discovered, analyzing, validated, rejected, archived]
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Number of items per page
        has_next:
          type: boolean
          description: Whether there are more items
        next_cursor:
          type: string
          description: Cursor for the next page
        prev_cursor:
          type: string
          description: Cursor for the previous page
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                description: Error code
                example: "VALIDATION_ERROR"
              message:
                type: string
                description: Human-readable error message
                example: "The provided email address is invalid"
              field:
                type: string
                description: Field that caused the error (if applicable)
                example: "email"
              details:
                type: object
                description: Additional error details
                additionalProperties: true
        meta:
          type: object
          properties:
            request_id:
              type: string
              format: uuid
              description: Unique request identifier for debugging
            timestamp:
              type: string
              format: date-time
              description: When the error occurred